fix : 이미지가 이미 null일 때 삭제하는 경우 반영

+++ run/src/main/java/com/guide/run/user/service/ImgService.java
@@ -44,6 +44,9 @@ public class ImgService {
     public void deleteProfile(String privateId){
         User user = userRepository.findUserByPrivateId(privateId).orElseThrow(NotExistUserException::new);
         String img = user.getImg();
+        if(img.isBlank()){
+            return;
+        }
         s3Uploader.deleteFile(img);
         user.editImg(null);
     }
